requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
Django==1.11.4
djangorestframework==3.6.3
h5py==2.7.0
html5lib==0.9999999
Keras==2.0.6
Markdown==2.6.8
numpy==1.13.1
olefile==0.44
peewee==2.10.1
Pillow==4.2.1
protobuf==3.3.0
pytz==2017.2
PyYAML==3.12
scipy==0.19.1
six==1.10.0
tensorflow==1.2.1
Theano==0.9.0
Werkzeug==0.12.2
